Remove plugin definition file and callback function move into class static member
